NETFLIX-BUILD: Update archive distribution to use public resources.
This calculates a checksum based on file contents and uses it to build a
persistent file path in HDFS. When those paths are added to the YARN
localizer as public, they shoudl be shared across applications.

Also includes the following commits:

4f3b804 NETFLIX-BUILD: Sort jars in distributed zip to get a consistent hash.
cef7d8a NETFLIX-BUILD: Set modified time from file in zip archive.
5e41145 NETFLIX-BUILD: Use a consistent file name for Spark assembly.
2e2ea9d NETFLIX-BUILD: Fix jar archive file path.