NETFLIX-BUILD: Add option to insert safe casts.
When casts are added, they are currently unsafe and silently allow
truncating fields. This adds an option, spark.sql.insertSafeCasts, that
changes the default behavior of insertInto so that only safe casts are
inserted. This is off by default to ensure unit tests pass, but is
turned on in the Netflix spark config.

(cherry picked from commit 92ded2c2ac894123775229deba30e257737d0b06)

Conflicts:
	netflix/spark-defaults.conf
	sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/Analyzer.scala