NETFLIX-BUILD: Add option to insert safe casts.
When casts are added, they are currently unsafe and silently allow
truncating fields. This adds an option, spark.sql.insertSafeCasts, that
changes the default behavior of insertInto so that only safe casts are
inserted. This is off by default to ensure unit tests pass, but is
turned on in the Netflix spark config.