NETFLIX-BUILD: Do not fail default location when checking for S3.
The saveAsTable method checks whether it is safe to use Hive's CTAS
logical plan node. Only S3 writes are supported, so this would
previously check whether the location was in S3. This skips failing the
write when the location isn't set and the table will use the default S3
location.