NETFLIX-BUILD: Copy hive-site.xml into jars/.
The hive-config application will ensure the correct hive-site.xml file
is in Genie's cluster configs directory. The remaining problem is that
drivers running in cluster mode won't see Genie's cluster config
directory and the hive-site.xml needs to be added to the driver's
classpath. The easiest way to add it is by shipping it with the jars
folder.