[SPARK-22160][SQL] Make sample points per partition (in range partitioner) configurable and bump the default value up to 100
Spark's RangePartitioner hard codes the number of sampling points per partition to be 20. This is sometimes too low. This ticket makes it configurable, via spark.sql.execution.rangeExchange.sampleSizePerPartition, and raises the default in Spark SQL to be 100.

Added a pretty sophisticated test based on chi square test ...

Author: Reynold Xin <rxin@databricks.com>

Closes #19387 from rxin/SPARK-22160.

This commit contains the following squashed commits:
938326b NETFLIX-BUILD: Fixup backport of SPARK-22160.