NETFLIX-BUILD: Add blacklist settings.
This enables the blacklist feature and updates the number of attempts
that will be run for a single task on an executor to 2. This will avoid
failing stages when there aren't enough executors for a bad task to
fail. For example, with a minimum of 3 executors by default, jobs with
errors that cause task failures will show the wrong error message:

"Aborting TaskSet 0.0 because task 0 (partition 0) cannot run anywhere
due to node and executor blacklist"

This is because the task needs to fail 4 times, but could only run once
on each executor. Bumping the max number of failures per executor to 2
avoids the placement problem by allowing up to 6 attmempts for each task
on 3 executors. Increasing the max failures for a task requires also
increasing max attempts per node so that executors on the same node will
be used and increasing max failed tasks per executor so that executors
aren't blacklisted because a single task failed.