SPARK-25006: Add CatalogTableIdentifier.
This adds CatalogTableIdentifier, which is an identifier that consists
of a triple: catalog, database, and table. Catalog and database are
optional.

The existing TableIdentifier class extends CatalogTableIdentifier and is
guarateed to have no defined catalog component. Classes that expect a
TableIdentifier should continue to use TableIdentifier to ensure that
catalogs are not leaked into code paths that do not support them.

(cherry picked from commit 00295ee6b3713995641c90a9b3b7cd4a6b79ded6)

Conflicts:
	sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/unresolved.scala
	sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/parser/AstBuilder.scala