NETFLIX-BUILD: Fix CREATE TABLE for iceberg partitions.
This allows users to create identity partitions for columns that match
the (col)_(transform) naming convention, like utc_hour. The create
statement can use utc_hour_identity to add an identity partition for the
utc_hour column. Previously, Spark would try to create a hour(utc)
partition for a non-existent utc column.

In addition, this adds a check for existence of the referenced column
when a transform name is matched. So utc_hour will check whether utc is
a column and only add the hour transform it it is.

This includes the following squashed commit:
566f8b9 NETFLIX-BUILD: Fix work-around for iceberg identity columns.