NETFLIX-BUILD: Fix output commit coordinator tracking.
The fix for SPARK-24589 was broken by the backport for SPARK-24552
because the attempt ID variable was reused, but changed in the latter
backport. This separates the two uses to ensure that the task attempt
number, not the TID is passed to the coordinator.

(cherry picked from commit 2d8b725c1a977a39223e77aa5f472cf0950d9c39)

Conflicts:
	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/WriteToDataSourceV2.scala