NETFLIX-BUILD: Remove special cases for Iceberg.
This fixes two special cases for v2 or Iceberg tables:
* ResolveRelations detected v2 tables and converted to AppendData
* HiveSessionCatalog detected Iceberg tables and used v2 relations

To fix resolution, this adds a v1 BaseRelation for v2 sources,
V2AsBaseRelation, that is created when a v2 provider is detected in the
DataSource's resolveRelation method. This passes v2 sources as v1
sources until they can be converted by the new v2 analyzer rule in
DataSourceV2Analysis.

To fix the special case for Iceberg tables, this adds a NetflixAnalysis
rule that converts MetastoreRelation with table_type "iceberg" to a
LogicalRelation wrapping a V2AsBaseRelation with the v2 relation. This
makes Iceberg tables look like other v2 tables when they are loaded from
the HiveSessionCatalog and will trigger the new DataSourceV2Analysis
rules.