NETFLIX-BUILD: Fix column resolution for Iceberg tables.
This runs the rule to replace MetastoreRelation with a v2 relation in
the HiveSessionCatalog. Ideally, this would be done as an analyzer rule,
but Spark runs column resolution before the Netflix rules that replace
the relation. Because columns are already resolved, they use the
metastore relation's IDs and not the new relation's IDs. The solution is
to ensure the Netflix rule is run first, by running it before returning
the metastore relation.