NETFLIX-BUILD: Implement CREATE TABLE with transforms.
To pass transforms, this adds two SQL logical plans for create and CTAS
that carry the options passed from the parser. V1 and V2 plans are
converted from the SQL plans.

This includes the following squashed commit:
2b736a810da NETFLIX-BUILD: Fix CREATE TABLE with transforms.

The apply transform rule was not working because of strange match
limitations in Scala. This also updates DataSourceStrategy to not match
v2 providers.

Conflicts:
	sql/catalyst/src/main/antlr4/org/apache/spark/sql/catalyst/parser/SqlBase.g4
	sql/core/src/main/scala/org/apache/spark/sql/execution/SparkSqlParser.scala
	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/DataSourceStrategy.scala
	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/DataSourceV2Analysis.scala