NETFLIX-BUILD: Update SNAPSHOT and MIGRATE to commit on the driver.
This avoids double commits from speculative tasks. Speculation cannot be
disabled on a per-job or per-stage basis, so the only way to have
speculation on by default and correct the behavior of SNAPSHOT and
MIGRATE is to perform table commits on the driver.

This uses an iterator to pull files back and commit groups of 500,000
files to avoid driver OOM errors.